+++ src/modules/MouseWithoutBorders/App/Class/Common.WinAPI.cs
@@ -269,7 +269,7 @@ namespace MouseWithoutBorders
             if (!Common.RunWithNoAdminRight)
             {
                 Logger.LogDebug("*** Starting on active Desktop: " + desktopToRunMouseWithoutBordersOn);
-                StartMouseWithoutBordersService(desktopToRunMouseWithoutBordersOn);
+                Service.StartMouseWithoutBordersService(desktopToRunMouseWithoutBordersOn);
             }
         }
 
@@ -279,7 +279,7 @@ namespace MouseWithoutBorders
             {
                 if (!IsMyDesktopActive() || Common.CurrentProcess.SessionId != NativeMethods.WTSGetActiveConsoleSessionId())
                 {
-                    Common.RunDDHelper(true);
+                    Helper.RunDDHelper(true);
                     int waitCount = 20;
 
                     while (NativeMethods.WTSGetActiveConsoleSessionId() == 0xFFFFFFFF && waitCount > 0)
