+++ src/modules/MouseWithoutBorders/MouseWithoutBorders.UnitTests/Core/Logger.PrivateDump.expected.txt
@@ -99,6 +99,111 @@ LegalKeyDictionary = Concurrent.ConcurrentDictionary`2[System.String,System.Byte
 --_budget = ????????????
 --_growLockArray = True
 --_comparerIsDefaultForClasses = False
+initDone = False
+REOPEN_WHEN_WSAECONNRESET = -10054
+REOPEN_WHEN_HOTKEY = -10055
+PleaseReopenSocket = 0
+ReopenSocketDueToReadError = False
+<LastResumeSuspendTime>k__BackingField = ????????????
+--_dateData = ????????????
+--MinValue = 01/01/0001 00:00:00
+--MaxValue = 31/12/9999 23:59:59
+--UnixEpoch = 01/01/1970 00:00:00
+lastReleaseAllKeysCall = 0
+PackageSent = MouseWithoutBorders.PackageMonitor
+--Keyboard = 0
+--Mouse = 0
+--Heartbeat = 0
+--ByeBye = 0
+--Hello = 0
+--Matrix = 0
+--ClipboardText = 0
+--ClipboardImage = 0
+--Clipboard = 0
+--ClipboardDragDrop = 0
+--ClipboardDragDropEnd = 0
+--ClipboardAsk = 0
+--ExplorerDragDrop = 0
+--Nil = 0
+PackageReceived = MouseWithoutBorders.PackageMonitor
+--Keyboard = 0
+--Mouse = 0
+--Heartbeat = 0
+--ByeBye = 0
+--Hello = 0
+--Matrix = 0
+--ClipboardText = 0
+--ClipboardImage = 0
+--Clipboard = 0
+--ClipboardDragDrop = 0
+--ClipboardDragDropEnd = 0
+--ClipboardAsk = 0
+--ExplorerDragDrop = 0
+--Nil = 0
+PackageID = 0
+SensitivePoints = Generic.List`1[Point]
+--_items = Point[]
+----System.Drawing.Point[] = Point[]: N/A
+--_size = 0
+--_version = 0
+--s_emptyArray = Point[]
+----System.Drawing.Point[] = Point[]: N/A
+p = {X=0,Y=0}
+--x = 0
+--y = 0
+--Empty = {X=0,Y=0}
+<IpcChannelCreated>k__BackingField = False
+BIG_CLIPBOARD_DATA_TIMEOUT = 30000
+MAX_CLIPBOARD_DATA_SIZE_CAN_BE_SENT_INSTANTLY_TCP = 1048576
+MAX_CLIPBOARD_FILE_SIZE_CAN_BE_SENT = 104857600
+TEXT_HEADER_SIZE = 12
+DATA_SIZE = 48
+TEXT_TYPE_SEP = {4CFF57F7-BEDD-43d5-AE8F-27A61E886F2F}
+TOGGLE_ICONS_SIZE = 4
+ICON_ONE = 0
+ICON_ALL = 1
+ICON_SMALL_CLIPBOARD = 2
+ICON_BIG_CLIPBOARD = 3
+ICON_ERROR = 4
+JUST_GOT_BACK_FROM_SCREEN_SAVER = 9999
+NETWORK_STREAM_BUF_SIZE = 1048576
+SymAlBlockSize = 16
+PW_LENGTH = 16
+PACKAGE_SIZE = 32
+PACKAGE_SIZE_EX = 64
+WP_PACKAGE_SIZE = 6
+KEYEVENTF_KEYDOWN = 1
+KEYEVENTF_KEYUP = 2
+WH_MOUSE = 7
+WH_KEYBOARD = 2
+WH_MOUSE_LL = 14
+WH_KEYBOARD_LL = 13
+WM_MOUSEMOVE = 512
+WM_LBUTTONDOWN = 513
+WM_RBUTTONDOWN = 516
+WM_MBUTTONDOWN = 519
+WM_XBUTTONDOWN = 523
+WM_LBUTTONUP = 514
+WM_RBUTTONUP = 517
+WM_MBUTTONUP = 520
+WM_XBUTTONUP = 524
+WM_LBUTTONDBLCLK = 515
+WM_RBUTTONDBLCLK = 518
+WM_MBUTTONDBLCLK = 521
+WM_MOUSEWHEEL = 522
+WM_KEYDOWN = 256
+WM_KEYUP = 257
+WM_SYSKEYDOWN = 260
+WM_SYSKEYUP = 261
+[DragDrop]
+===============
+isDragging = False
+dragDropStep05ExCalledByIpc = 0
+isDropping = False
+dragMachine = NONE
+<MouseDown>k__BackingField = False
+[Event]
+===============
 KeybdPackage = MouseWithoutBorders.DATA
 --Type = 0
 --Id = 0
@@ -144,113 +249,15 @@ actualLastPos = {X=0,Y=0}
 --Empty = {X=0,Y=0}
 myLastX = 0
 myLastY = 0
+[Helper]
+===============
 signalHelperToExit = False
 signalWatchDogToExit = False
 WndProcCounter = 0
-initDone = False
-REOPEN_WHEN_WSAECONNRESET = -10054
-REOPEN_WHEN_HOTKEY = -10055
-PleaseReopenSocket = 0
-ReopenSocketDueToReadError = False
-<LastResumeSuspendTime>k__BackingField = ????????????
---_dateData = ????????????
---MinValue = 01/01/0001 00:00:00
---MaxValue = 31/12/9999 23:59:59
---UnixEpoch = 01/01/1970 00:00:00
-lastReleaseAllKeysCall = 0
-PackageSent = MouseWithoutBorders.PackageMonitor
---Keyboard = 0
---Mouse = 0
---Heartbeat = 0
---ByeBye = 0
---Hello = 0
---Matrix = 0
---ClipboardText = 0
---ClipboardImage = 0
---Clipboard = 0
---ClipboardDragDrop = 0
---ClipboardDragDropEnd = 0
---ClipboardAsk = 0
---ExplorerDragDrop = 0
---Nil = 0
-PackageReceived = MouseWithoutBorders.PackageMonitor
---Keyboard = 0
---Mouse = 0
---Heartbeat = 0
---ByeBye = 0
---Hello = 0
---Matrix = 0
---ClipboardText = 0
---ClipboardImage = 0
---Clipboard = 0
---ClipboardDragDrop = 0
---ClipboardDragDropEnd = 0
---ClipboardAsk = 0
---ExplorerDragDrop = 0
---Nil = 0
-PackageID = 0
-shownErrMessage = False
-lastStartServiceTime = ????????????
---_dateData = ????????????
---MinValue = 01/01/0001 00:00:00
---MaxValue = 31/12/9999 23:59:59
---UnixEpoch = 01/01/1970 00:00:00
-SensitivePoints = Generic.List`1[Point]
---_items = Point[]
-----System.Drawing.Point[] = Point[]: N/A
---_size = 0
---_version = 0
---s_emptyArray = Point[]
-----System.Drawing.Point[] = Point[]: N/A
-p = {X=0,Y=0}
---x = 0
---y = 0
---Empty = {X=0,Y=0}
-<IpcChannelCreated>k__BackingField = False
-BIG_CLIPBOARD_DATA_TIMEOUT = 30000
-MAX_CLIPBOARD_DATA_SIZE_CAN_BE_SENT_INSTANTLY_TCP = 1048576
-MAX_CLIPBOARD_FILE_SIZE_CAN_BE_SENT = 104857600
-TEXT_HEADER_SIZE = 12
-DATA_SIZE = 48
-TEXT_TYPE_SEP = {4CFF57F7-BEDD-43d5-AE8F-27A61E886F2F}
-TOGGLE_ICONS_SIZE = 4
-ICON_ONE = 0
-ICON_ALL = 1
-ICON_SMALL_CLIPBOARD = 2
-ICON_BIG_CLIPBOARD = 3
-ICON_ERROR = 4
-JUST_GOT_BACK_FROM_SCREEN_SAVER = 9999
-NETWORK_STREAM_BUF_SIZE = 1048576
-SymAlBlockSize = 16
-PW_LENGTH = 16
 HELPER_FORM_TEXT = Mouse without Borders Helper
 HelperProcessName = PowerToys.MouseWithoutBordersHelper
-PACKAGE_SIZE = 32
-PACKAGE_SIZE_EX = 64
-WP_PACKAGE_SIZE = 6
-KEYEVENTF_KEYDOWN = 1
-KEYEVENTF_KEYUP = 2
-WH_MOUSE = 7
-WH_KEYBOARD = 2
-WH_MOUSE_LL = 14
-WH_KEYBOARD_LL = 13
-WM_MOUSEMOVE = 512
-WM_LBUTTONDOWN = 513
-WM_RBUTTONDOWN = 516
-WM_MBUTTONDOWN = 519
-WM_XBUTTONDOWN = 523
-WM_LBUTTONUP = 514
-WM_RBUTTONUP = 517
-WM_MBUTTONUP = 520
-WM_XBUTTONUP = 524
-WM_LBUTTONDBLCLK = 515
-WM_RBUTTONDBLCLK = 518
-WM_MBUTTONDBLCLK = 521
-WM_MOUSEWHEEL = 522
-WM_KEYDOWN = 256
-WM_KEYUP = 257
-WM_SYSKEYDOWN = 260
-WM_SYSKEYUP = 261
+[Launch]
+===============
 [Logger]
 ===============
 AllLogsLock = Lock
@@ -319,13 +326,6 @@ MAX_LOG = 10000
 MaxLogExceptionPerHour = 1000
 HeaderSENT = Be{0},Ke{1},Mo{2},He{3},Mx{4},Tx{5},Im{6},By{7},Cl{8},Dr{9},De{10},Ed{11},Ie{12},Ni{13}
 HeaderRECEIVED = Be{0},Ke{1},Mo{2},He{3},Mx{4},Tx{5},Im{6},By{7},Cl{8},Dr{9},De{10},Ed{11},In{12},Ni{13},Pc{14}/{15}
-[DragDrop]
-===============
-isDragging = False
-dragDropStep05ExCalledByIpc = 0
-isDropping = False
-dragMachine = NONE
-<MouseDown>k__BackingField = False
 [MachineStuff]
 ===============
 McMatrixLock = Lock
@@ -424,3 +424,11 @@ lastXY = {X=0,Y=0}
 --x = 0
 --y = 0
 --Empty = {X=0,Y=0}
+[Service]
+===============
+shownErrMessage = False
+lastStartServiceTime = ????????????
+--_dateData = ????????????
+--MinValue = 01/01/0001 00:00:00
+--MaxValue = 31/12/9999 23:59:59
+--UnixEpoch = 01/01/1970 00:00:00
